backend/requirements.txt
# Core Django
Django==5.0.9
psycopg2-binary==2.9.9
python-dotenv==1.0.1
django-environ==0.11.2

# Django REST Framework
djangorestframework==3.15.2
drf-spectacular==0.27.2
django-filter==24.3
django-cors-headers==4.4.0
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1

# Channels & WebSocket (Real-time chat)
channels==4.1.0
channels-redis==4.2.0
daphne==4.1.2
redis==5.0.8

# Database adapter for PostgreSQL
dj-database-url==2.2.0

# Static & Media files
whitenoise==6.7.0
Pillow==10.4.0

# Security
python-decouple==3.8
cryptography==43.0.0
